+++ src/open_sstv/ui/main_window.py
@@ -463,6 +463,7 @@ class MainWindow(QMainWindow):
         self._input_device = new_input
         self._rx_worker.set_input_gain(self._config.audio_input_gain)
         self._rx_worker.set_weak_signal(self._config.rx_weak_signal_mode)
+        self._rx_worker.set_final_slant_correction(self._config.apply_final_slant_correction)
         self._tx_worker.set_cw_id(
             self._config.cw_id_enabled,
             self._config.callsign,